+++ js/get-random-array-element.js
@@ -1,0 +1,7 @@
+import { getRandomInteger } from './get-random-element.js';
+
+// функция для получения случайного элемента массива
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomArrayElement };
